refactor Epmc move to folder epmc refactor cooc move to File EpmcCooccurrences refactor uri configured in reference.conf fix output format
