+++ src/main/resources/reference.conf
@@ -77,16 +77,18 @@ common {
 }
 
 epmc {
+  uris {
+    ensembl = "https://www.ensembl.org/Homo_sapiens/Gene/Summary?db=core;g="
+    chembl = "https://www.ebi.ac.uk/chembl/compound_report_card/"
+    ontologies = "https://www.ebi.ac.uk/ols/ontologies/efo/terms?short_form="
+  }
   input {
     cooccurences = ${literature.processing.outputs.cooccurrences}
   }
   outputs {
     epmc-cooccurrences {
-      format = "json" // to be consistent with the other Evidence files
+      format = ${common.output-format}
       path = ${common.output}"/epmc/cooccurrences"
-      options = [
-        {k: "compression", v: "gzip"},
-      ]
     }
     output {
       format = "json" // to be consistent with the other Evidence files
